+++ app_plus/src/main/java/org/gkisalatiga/plus/fragment/FragmentServices.kt
@@ -40,7 +40,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
@@ -50,6 +49,7 @@ import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import org.gkisalatiga.plus.R
 import org.gkisalatiga.plus.composable.YouTubeViewCompanion
+import org.gkisalatiga.plus.data.ActivityData
 import org.gkisalatiga.plus.db.MainCompanion
 import org.gkisalatiga.plus.global.GlobalCompanion
 import org.gkisalatiga.plus.lib.AppNavigation
@@ -61,7 +61,7 @@ import org.gkisalatiga.plus.screen.ScreenVideoListCompanion
 import org.json.JSONArray
 import org.json.JSONObject
 
-class FragmentServices : ComponentActivity() {
+class FragmentServices (private val current : ActivityData) : ComponentActivity() {
 
     @Composable
     fun getComposable() {
@@ -158,7 +158,6 @@ class FragmentServices : ComponentActivity() {
         }
 
         /* Displaying the individual item of this section. */
-        val ctx = LocalContext.current
         LazyRow {
             // Left-right padding.
             item { Spacer(modifier = Modifier.width(5.dp)) }
@@ -179,7 +178,7 @@ class FragmentServices : ComponentActivity() {
 
                 Card (
                     onClick = {
-                        if (GlobalCompanion.DEBUG_ENABLE_TOAST) Toast.makeText(ctx, "The card $title is clicked", Toast.LENGTH_SHORT).show()
+                        if (GlobalCompanion.DEBUG_ENABLE_TOAST) Toast.makeText(current.ctx, "The card $title is clicked", Toast.LENGTH_SHORT).show()
 
                         // Trying to switch to the YouTube viewer and open the stream.
                         Logger.log({}, "Opening the YouTube stream: $url.")