+++ app_plus/src/main/java/org/gkisalatiga/plus/screen/ScreenGaleriYear.kt
@@ -38,7 +38,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextOverflow
@@ -47,6 +46,7 @@ import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import org.gkisalatiga.plus.R
 import org.gkisalatiga.plus.composable.TopAppBarColorScheme
+import org.gkisalatiga.plus.data.ActivityData
 import org.gkisalatiga.plus.db.GalleryCompanion
 import org.gkisalatiga.plus.global.GlobalCompanion
 import org.gkisalatiga.plus.lib.AppNavigation
@@ -56,7 +56,7 @@ import org.gkisalatiga.plus.lib.StringFormatter
 import org.json.JSONArray
 import org.json.JSONObject
 
-class ScreenGaleriYear : ComponentActivity() {
+class ScreenGaleriYear (private val current : ActivityData) : ComponentActivity() {
 
     @Composable
     @SuppressLint("ComposableNaming", "UnusedMaterial3ScaffoldPaddingParameter")
@@ -81,7 +81,6 @@ class ScreenGaleriYear : ComponentActivity() {
     @Suppress("RemoveCurlyBracesFromTemplate")
     @SuppressLint("ComposableNaming")
     private fun getMainContent() {
-        val ctx = LocalContext.current
 
         // The agenda node.
         val galleryNode = GalleryCompanion.jsonRoot!!
@@ -130,7 +129,7 @@ class ScreenGaleriYear : ComponentActivity() {
                     // Displaying the individual card.
                     Card(
                         onClick = {
-                            if (GlobalCompanion.DEBUG_ENABLE_TOAST) Toast.makeText(ctx, "Opening gallery album year: $title", Toast.LENGTH_SHORT).show()
+                            if (GlobalCompanion.DEBUG_ENABLE_TOAST) Toast.makeText(current.ctx, "Opening gallery album year: $title", Toast.LENGTH_SHORT).show()
 
                             // Navigate to the WebView viewer.
                             ScreenGaleriListCompanion.putArguments(