+++ app_plus/src/main/java/org/gkisalatiga/plus/screen/ScreenPersembahan.kt
@@ -50,7 +50,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
@@ -61,12 +60,13 @@ import androidx.compose.ui.unit.sp
 import coil.compose.AsyncImage
 import org.gkisalatiga.plus.R
 import org.gkisalatiga.plus.composable.TopAppBarColorScheme
+import org.gkisalatiga.plus.data.ActivityData
 import org.gkisalatiga.plus.db.MainCompanion
 import org.gkisalatiga.plus.lib.AppNavigation
 import org.gkisalatiga.plus.services.ClipManager
 import org.json.JSONObject
 
-class ScreenPersembahan : ComponentActivity() {
+class ScreenPersembahan (private val current : ActivityData) : ComponentActivity() {
     private var isFirstElement = false
     private val showOffertoryCodeTextDialog = mutableStateOf(false)
 
@@ -94,7 +94,6 @@ class ScreenPersembahan : ComponentActivity() {
 
     @Composable
     private fun getMainContent() {
-        val ctx = LocalContext.current
 
         // The column's saved scroll state.
         val scrollState = ScreenPersembahanCompanion.rememberedScrollState!!
@@ -176,7 +175,7 @@ class ScreenPersembahan : ComponentActivity() {
                         val clipData = ClipData.newPlainText("text", currentNode.getString("bank-number").replace(".", ""))
                         ClipManager.clipManager!!.setPrimaryClip(clipData)
 
-                        Toast.makeText(ctx, notificationString, Toast.LENGTH_SHORT).show()
+                        Toast.makeText(current.ctx, notificationString, Toast.LENGTH_SHORT).show()
                     })
                 )
                 HorizontalDivider()
@@ -222,7 +221,7 @@ class ScreenPersembahan : ComponentActivity() {
                         val clipData = ClipData.newPlainText("text", currentNode.getString("unique-code"))
                         ClipManager.clipManager!!.setPrimaryClip(clipData)
 
-                        Toast.makeText(ctx, notificationString, Toast.LENGTH_SHORT).show()
+                        Toast.makeText(current.ctx, notificationString, Toast.LENGTH_SHORT).show()
                     })
                 )
                 HorizontalDivider()